backend/requirements.txt
# Core FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Data Processing & Analysis
pandas==2.1.4
numpy>=1.26.0
scipy>=1.11.4

# Financial Data & Market Data
yfinance>=0.2.28
alpha-vantage==2.3.1

# Alternative Data Sources
newsapi-python==0.2.7
praw==7.7.1
requests==2.31.0

# NLP & Sentiment Analysis
textblob==0.17.1
nltk==3.8.1

# Machine Learning
xgboost==2.0.3
scikit-learn==1.4.0
joblib==1.3.2

# Statistical Analysis
statsmodels==0.14.1

# C++ Integration (optional)
pybind11==2.11.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
tqdm==4.66.1
pytz==2024.1